Move task item-related components to TaskItem subdirectory for organization
